Upgrade to @sentry/nextjs 8.13.0
The configuration for the server needs to be imported from a next.js instrumentation file `instrumentation.js|ts`. As we're not using edge runtime, neither `sentry.edge.config` or an additional branch in `instrumentation.js` is needed.

Setting a `sentry` property on the Next.js config object as a means of configuration is no longer supported. The `sentryBuildOptions` argument of the `withSentryConfig()` function instead must be used instead.